pyproject.toml
[project]
name = "dolly-cam"
version = "0.1.0"
description = "Touchscreen-controlled Raspberry Pi webcam recorder with Google Drive sync."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client",
    "google-auth",
    "google-auth-httplib2",
    "tomli; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
]

[project.scripts]
dolly-cam = "dolly_cam.app:main"

[tool.setuptools.packages.find]
where = ["src"]
